+++ CHANGELOG.md
@@ -7,7 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.1.14] - Unreleased
-No changes yet
+## Changed
+- Use MW_transp=1 when the fiberassign FAPRGRM is "BACKUP". Fixes [#8](https://github.com/desihub/desietc/issues/8).
 
 ## [0.1.13] - 2021-10-07
 ## Added